+++ inst/tinytest/test_helpers.R
@@ -4,3 +4,7 @@ expect_error(alkahest:::assert_odd(x))
 
 x <- 3
 expect_equal(alkahest:::assert_odd(x), x)
+
+# Length =======================================================================
+expect_error(alkahest:::assert_length(1:5, 3))
+expect_equal(alkahest:::assert_length(1:5, 5), 1:5)