+++ R/alkahest-package.R
@@ -2,7 +2,7 @@
 #' \tabular{ll}{
 #'  **Package:** \tab alkahest \cr
 #'  **Type:** \tab Package \cr
-#'  **Version:** \tab 1.1.1 \cr
+#'  **Version:** \tab 1.2.0 \cr
 #'  **License:** \tab GPL-3 \cr
 #'  **Zenodo:** \tab \doi{10.5281/zenodo.7081524} \cr
 #' }