+++ NEWS.md
@@ -1,4 +1,6 @@
 # alkahest 1.1.1.9000
+## New classes and methods
+* Add baseline estimation with asymmetric least squares smoothing (Eilers and Boelens, 2005).
 
 # alkahest 1.1.1
 ## Bugfixes & changes